+++ db/hints/internal/hint_endpoint_manager.cc
@@ -296,7 +296,7 @@ public:
     no_column_mapping(const table_schema_version& id) : std::out_of_range(format("column mapping for CF schema_version {} is missing", id)) {}
 };
 
-future<> end_point_hints_manager::sender::flush_maybe() noexcept {
+future<> sender::flush_maybe() noexcept {
     auto current_time = clock::now();
     if (current_time >= _next_flush_tp) {
         return _ep_manager.flush_current_hints().then([this, current_time] {
@@ -309,7 +309,7 @@ future<> end_point_hints_manager::sender::flush_maybe() noexcept {
     return make_ready_future<>();
 }
 
-future<timespec> end_point_hints_manager::sender::get_last_file_modification(const sstring& fname) {
+future<timespec> sender::get_last_file_modification(const sstring& fname) {
     return open_file_dma(fname, open_flags::ro).then([] (file f) {
         return do_with(std::move(f), [] (file& f) {
             return f.stat();
@@ -319,7 +319,7 @@ future<timespec> end_point_hints_manager::sender::get_last_file_modification(con
     });
 }
 
-future<> end_point_hints_manager::sender::do_send_one_mutation(frozen_mutation_and_schema m, locator::effective_replication_map_ptr ermp, const inet_address_vector_replica_set& natural_endpoints) noexcept {
+future<> sender::do_send_one_mutation(frozen_mutation_and_schema m, locator::effective_replication_map_ptr ermp, const inet_address_vector_replica_set& natural_endpoints) noexcept {
     return futurize_invoke([this, m = std::move(m), ermp = std::move(ermp), &natural_endpoints] () mutable -> future<> {
         // The fact that we send with CL::ALL in both cases below ensures that new hints are not going
         // to be generated as a result of hints sending.
@@ -333,7 +333,7 @@ future<> end_point_hints_manager::sender::do_send_one_mutation(frozen_mutation_a
     });
 }
 
-bool end_point_hints_manager::sender::can_send() noexcept {
+bool sender::can_send() noexcept {
     if (stopping() && !draining()) {
         return false;
     }
@@ -354,7 +354,7 @@ bool end_point_hints_manager::sender::can_send() noexcept {
     }
 }
 
-frozen_mutation_and_schema end_point_hints_manager::sender::get_mutation(lw_shared_ptr<send_one_file_ctx> ctx_ptr, fragmented_temporary_buffer& buf) {
+frozen_mutation_and_schema sender::get_mutation(lw_shared_ptr<send_one_file_ctx> ctx_ptr, fragmented_temporary_buffer& buf) {
     hint_entry_reader hr(buf);
     auto& fm = hr.mutation();
     auto& cm = get_column_mapping(std::move(ctx_ptr), fm, hr);
@@ -369,7 +369,7 @@ frozen_mutation_and_schema end_point_hints_manager::sender::get_mutation(lw_shar
     return {std::move(hr).mutation(), std::move(schema)};
 }
 
-const column_mapping& end_point_hints_manager::sender::get_column_mapping(lw_shared_ptr<send_one_file_ctx> ctx_ptr, const frozen_mutation& fm, const hint_entry_reader& hr) {
+const column_mapping& sender::get_column_mapping(lw_shared_ptr<send_one_file_ctx> ctx_ptr, const frozen_mutation& fm, const hint_entry_reader& hr) {
     auto cm_it = ctx_ptr->schema_ver_to_column_mapping.find(fm.schema_version());
     if (cm_it == ctx_ptr->schema_ver_to_column_mapping.end()) {
         if (!hr.get_column_mapping()) {
@@ -383,7 +383,7 @@ const column_mapping& end_point_hints_manager::sender::get_column_mapping(lw_sha
     return cm_it->second;
 }
 
-end_point_hints_manager::sender::sender(end_point_hints_manager& parent, service::storage_proxy& local_storage_proxy,replica::database& local_db, gms::gossiper& local_gossiper) noexcept
+sender::sender(end_point_hints_manager& parent, service::storage_proxy& local_storage_proxy,replica::database& local_db, gms::gossiper& local_gossiper) noexcept
     : _stopped(make_ready_future<>())
     , _ep_key(parent.end_point_key())
     , _ep_manager(parent)
@@ -396,7 +396,7 @@ end_point_hints_manager::sender::sender(end_point_hints_manager& parent, service
     , _file_update_mutex(_ep_manager.file_update_mutex())
 {}
 
-end_point_hints_manager::sender::sender(const sender& other, end_point_hints_manager& parent) noexcept
+sender::sender(const sender& other, end_point_hints_manager& parent) noexcept
     : _stopped(make_ready_future<>())
     , _ep_key(parent.end_point_key())
     , _ep_manager(parent)
@@ -409,12 +409,12 @@ end_point_hints_manager::sender::sender(const sender& other, end_point_hints_man
     , _file_update_mutex(_ep_manager.file_update_mutex())
 {}
 
-end_point_hints_manager::sender::~sender() {
+sender::~sender() {
     dismiss_replay_waiters();
 }
 
 
-future<> end_point_hints_manager::sender::stop(drain should_drain) noexcept {
+future<> sender::stop(drain should_drain) noexcept {
     return seastar::async([this, should_drain] {
         set_stopping();
         _stop_as.request_abort();
@@ -449,15 +449,15 @@ future<> end_point_hints_manager::sender::stop(drain should_drain) noexcept {
     });
 }
 
-void end_point_hints_manager::sender::add_segment(sstring seg_name) {
+void sender::add_segment(sstring seg_name) {
     _segments_to_replay.emplace_back(std::move(seg_name));
 }
 
-void end_point_hints_manager::sender::add_foreign_segment(sstring seg_name) {
+void sender::add_foreign_segment(sstring seg_name) {
     _foreign_segments_to_replay.emplace_back(std::move(seg_name));
 }
 
-end_point_hints_manager::sender::clock::duration end_point_hints_manager::sender::next_sleep_duration() const {
+sender::clock::duration sender::next_sleep_duration() const {
     clock::time_point current_time = clock::now();
     clock::time_point next_flush_tp = std::max(_next_flush_tp, current_time);
     clock::time_point next_retry_tp = std::max(_next_send_retry_tp, current_time);
@@ -468,7 +468,7 @@ end_point_hints_manager::sender::clock::duration end_point_hints_manager::sender
     return clock::duration(10 * div_ceil(d.count(), 10));
 }
 
-void end_point_hints_manager::sender::start() {
+void sender::start() {
     seastar::thread_attributes attr;
 
     attr.sched_group = _hints_cpu_sched_group;
@@ -492,7 +492,7 @@ void end_point_hints_manager::sender::start() {
     });
 }
 
-future<> end_point_hints_manager::sender::send_one_mutation(frozen_mutation_and_schema m) {
+future<> sender::send_one_mutation(frozen_mutation_and_schema m) {
     auto erm = _db.find_column_family(m.s).get_effective_replication_map();
     auto token = dht::get_token(*m.s, m.fm.key());
     inet_address_vector_replica_set natural_endpoints = erm->get_natural_endpoints(std::move(token));
@@ -500,7 +500,7 @@ future<> end_point_hints_manager::sender::send_one_mutation(frozen_mutation_and_
     return do_send_one_mutation(std::move(m), std::move(erm), std::move(natural_endpoints));
 }
 
-future<> end_point_hints_manager::sender::send_one_hint(lw_shared_ptr<send_one_file_ctx> ctx_ptr, fragmented_temporary_buffer buf, db::replay_position rp, gc_clock::duration secs_since_file_mod, const sstring& fname) {
+future<> sender::send_one_hint(lw_shared_ptr<send_one_file_ctx> ctx_ptr, fragmented_temporary_buffer buf, db::replay_position rp, gc_clock::duration secs_since_file_mod, const sstring& fname) {
     return _resource_manager.get_send_units_for(buf.size_bytes()).then([this, secs_since_file_mod, &fname, buf = std::move(buf), rp, ctx_ptr] (auto units) mutable {
         ctx_ptr->mark_hint_as_in_progress(rp);
 
@@ -570,7 +570,7 @@ future<> end_point_hints_manager::sender::send_one_hint(lw_shared_ptr<send_one_f
     });
 }
 
-void end_point_hints_manager::sender::notify_replay_waiters() noexcept {
+void sender::notify_replay_waiters() noexcept {
     if (!_foreign_segments_to_replay.empty()) {
         manager_logger.trace("[{}] notify_replay_waiters(): not notifying because there are still {} foreign segments to replay", end_point_key(), _foreign_segments_to_replay.size());
         return;
@@ -586,7 +586,7 @@ void end_point_hints_manager::sender::notify_replay_waiters() noexcept {
     }
 }
 
-void end_point_hints_manager::sender::dismiss_replay_waiters() noexcept {
+void sender::dismiss_replay_waiters() noexcept {
     for (auto& p : _replay_waiters) {
         manager_logger.debug("[{}] dismiss_replay_waiters(): dismissing one", end_point_key());
         auto ptr = p.second;
@@ -596,7 +596,7 @@ void end_point_hints_manager::sender::dismiss_replay_waiters() noexcept {
     _replay_waiters.clear();
 }
 
-future<> end_point_hints_manager::sender::wait_until_hints_are_replayed_up_to(abort_source& as, db::replay_position up_to_rp) {
+future<> sender::wait_until_hints_are_replayed_up_to(abort_source& as, db::replay_position up_to_rp) {
     manager_logger.debug("[{}] wait_until_hints_are_replayed_up_to(): entering with target {}", end_point_key(), up_to_rp);
     if (_foreign_segments_to_replay.empty() && up_to_rp < _sent_upper_bound_rp) {
         manager_logger.debug("[{}] wait_until_hints_are_replayed_up_to(): hints were already replayed above the point ({} < {})", end_point_key(), up_to_rp, _sent_upper_bound_rp);
@@ -628,18 +628,18 @@ future<> end_point_hints_manager::sender::wait_until_hints_are_replayed_up_to(ab
     });
 }
 
-void end_point_hints_manager::sender::send_one_file_ctx::mark_hint_as_in_progress(db::replay_position rp) {
+void sender::send_one_file_ctx::mark_hint_as_in_progress(db::replay_position rp) {
     in_progress_rps.insert(rp);
 }
 
-void end_point_hints_manager::sender::send_one_file_ctx::on_hint_send_success(db::replay_position rp) noexcept {
+void sender::send_one_file_ctx::on_hint_send_success(db::replay_position rp) noexcept {
     in_progress_rps.erase(rp);
     if (!last_succeeded_rp || *last_succeeded_rp < rp) {
         last_succeeded_rp = rp;
     }
 }
 
-void end_point_hints_manager::sender::send_one_file_ctx::on_hint_send_failure(db::replay_position rp) noexcept {
+void sender::send_one_file_ctx::on_hint_send_failure(db::replay_position rp) noexcept {
     in_progress_rps.erase(rp);
     segment_replay_failed = true;
     if (!first_failed_rp || rp < *first_failed_rp) {
@@ -647,7 +647,7 @@ void end_point_hints_manager::sender::send_one_file_ctx::on_hint_send_failure(db
     }
 }
 
-db::replay_position end_point_hints_manager::sender::send_one_file_ctx::get_replayed_bound() const noexcept {
+db::replay_position sender::send_one_file_ctx::get_replayed_bound() const noexcept {
     // We are sure that all hints were sent _below_ the position which is the minimum of the following:
     // - Position of the first hint that failed to be sent in this replay (first_failed_rp),
     // - Position of the last hint which was successfully sent (last_succeeded_rp, inclusive bound),
@@ -670,13 +670,13 @@ db::replay_position end_point_hints_manager::sender::send_one_file_ctx::get_repl
     return rp;
 }
 
-void end_point_hints_manager::sender::rewind_sent_replay_position_to(db::replay_position rp) {
+void sender::rewind_sent_replay_position_to(db::replay_position rp) {
     _sent_upper_bound_rp = rp;
     notify_replay_waiters();
 }
 
 // runs in a seastar::async context
-bool end_point_hints_manager::sender::send_one_file(const sstring& fname) {
+bool sender::send_one_file(const sstring& fname) {
     timespec last_mod = get_last_file_modification(fname).get0();
     gc_clock::duration secs_since_file_mod = std::chrono::seconds(last_mod.tv_sec);
     lw_shared_ptr<send_one_file_ctx> ctx_ptr = make_lw_shared<send_one_file_ctx>(_last_schema_ver_to_column_mapping);
@@ -760,7 +760,7 @@ bool end_point_hints_manager::sender::send_one_file(const sstring& fname) {
     return true;
 }
 
-const sstring* end_point_hints_manager::sender::name_of_current_segment() const {
+const sstring* sender::name_of_current_segment() const {
     // Foreign segments are replayed first
     if (!_foreign_segments_to_replay.empty()) {
         return &_foreign_segments_to_replay.front();
@@ -771,7 +771,7 @@ const sstring* end_point_hints_manager::sender::name_of_current_segment() const 
     return nullptr;
 }
 
-void end_point_hints_manager::sender::pop_current_segment() {
+void sender::pop_current_segment() {
     if (!_foreign_segments_to_replay.empty()) {
         _foreign_segments_to_replay.pop_front();
     } else if (!_segments_to_replay.empty()) {
@@ -780,7 +780,7 @@ void end_point_hints_manager::sender::pop_current_segment() {
 }
 
 // Runs in the seastar::async context
-void end_point_hints_manager::sender::send_hints_maybe() noexcept {
+void sender::send_hints_maybe() noexcept {
     using namespace std::literals::chrono_literals;
     manager_logger.trace("send_hints(): going to send hints to {}, we have {} segment to replay", end_point_key(), _segments_to_replay.size() + _foreign_segments_to_replay.size());
 
@@ -818,7 +818,7 @@ void end_point_hints_manager::sender::send_hints_maybe() noexcept {
     manager_logger.trace("send_hints(): we handled {} segments", replayed_segments_count);
 }
 
-hint_stats& end_point_hints_manager::sender::shard_stats() {
+hint_stats& sender::shard_stats() {
     return _shard_manager._stats;
 }
 
@@ -834,7 +834,7 @@ resource_manager& end_point_hints_manager::shard_resource_manager() {
     return _shard_manager._resource_manager;
 }
 
-bool end_point_hints_manager::sender::replay_allowed() const noexcept {
+bool sender::replay_allowed() const noexcept {
     return _ep_manager.replay_allowed();
 }
