+++ db/hints/manager.hh
@@ -83,6 +83,7 @@ private:
 
     friend class space_watchdog;
     friend class internal::end_point_hints_manager;
+    friend class internal::sender;
 
     using end_point_hints_manager = internal::end_point_hints_manager;
